fix event models
